Refactor services, core components, transports, managers
