+++ src/core/eventd/index.js
@@ -1,0 +1,12 @@
+const debug = require('debug')('canvas:eventd');
+
+class EventD {
+
+    constructor(options = {}) {
+        debug('Initializing Canvas EventD');
+        debug('Options:', options);
+    }
+
+}
+
+module.exports = EventD;
